renamed config.js to constants.js
