+++ features/about/oraganisms/About/index.js
@@ -3,8 +3,8 @@ import Descriptor from '@site/features/descriptor'
 import Slider from '@site/features/slider'
 import TextBlock from '@site/ui/molecules/TextBlock'
 import { ReactIcon, NodeIcon, SwiftIcon } from '@site/ui/atoms/icons'
+import { NON_BREAKING_SPACE } from '@site/constants'
 import AboutText from '../../atoms/AboutText'
-import { NON_BREAKING_SPACE } from '@site/config'
 import styles from './about.css'
 
 const About = () => (
